Create utils and api files, also a <Scenario /> component
