+++ src/components/index.js
@@ -1,3 +1,4 @@
+export * from "./Scenario"
 export * from "./Counter"
 export * from "./HP"
 export * from "./Monsters"
